adding tests in process editing

+++ src/com/epam/liavitskaya/main/service/impl/LibraryServiceImpl.java
@@ -1,5 +1,6 @@
 package com.epam.liavitskaya.main.service.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import com.epam.liavitskaya.main.bean.Book;
@@ -9,16 +10,18 @@ import com.epam.liavitskaya.main.dao.impl.SQLBookDao;
 import com.epam.liavitskaya.main.enums.BookStatus;
 import com.epam.liavitskaya.main.service.LibraryService;
 import com.epam.liavitskaya.main.service.exception.ServiceException;
-import com.epam.liavitskaya.main.utils.RequestParserUtil;
+import com.epam.liavitskaya.main.util.RequestParserUtil;
 
 public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public void addNewBookService(String request) throws ServiceException {
+
 		BookDAO bookDAO = new SQLBookDao();
-		String[] parseRequest = RequestParserUtil.parseRequest(request, 4);
+		String[] splitRequest = RequestParserUtil.parseRequest(request, 4);
+
 		try {
-			bookDAO.addBook(new Book(parseRequest[1], parseRequest[2], parseRequest[3], BookStatus.AVAILABLE));
+			bookDAO.addBook(new Book(splitRequest[1], splitRequest[2], splitRequest[3], BookStatus.AVAILABLE));
 		} catch (DAOException e) {
 			throw new ServiceException();
 		}
@@ -26,8 +29,13 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public void addEditedBookService(String request) throws ServiceException {
+
 		BookDAO bookDAO = new SQLBookDao();
 		try {
+			// int bookCount = bookDAO.rowCount();
+			// if (id < 1 || bookCount < id) {
+			// throw new ServiceException("incorrect id");
+			// }
 			bookDAO.editBook(new Book());
 		} catch (DAOException e) {
 			throw new ServiceException();
@@ -36,8 +44,10 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public List<Book> bookFondReviewService() throws ServiceException {
+
 		BookDAO bookDAO = new SQLBookDao();
-		List<Book> bookFondReview;
+		List<Book> bookFondReview = new ArrayList<>();
+
 		try {
 			bookFondReview = bookDAO.bookFondReview();
 		} catch (DAOException e) {
@@ -48,10 +58,16 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public void orderBookService(String request) throws ServiceException {
-		String idStr = request.substring(request.indexOf(" ") + 1, request.length());	
-		int id = Integer.parseInt(idStr);		
+
 		BookDAO bookDAO = new SQLBookDao();
+		String idStr = request.substring(request.indexOf(" ") + 1, request.length());
+
 		try {
+			int id = Integer.parseInt(idStr);
+			int bookCount = bookDAO.rowCount();
+			if (id < 1 || bookCount < id) {
+				throw new ServiceException("incorrect id");
+			}
 			bookDAO.changeBookStatus(BookStatus.ORDERED, id);
 		} catch (DAOException e) {
 			throw new ServiceException();
@@ -60,11 +76,15 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public void editBookDescriptionService(String request) throws ServiceException {
-		String idStr = request.substring(request.indexOf(" ") + 1 , request.length());
-		int id = Integer.parseInt(idStr);
-		System.out.println(id);
+
 		BookDAO bookDAO = new SQLBookDao();
-		Book book = new Book();          
+		String idStr = request.substring(request.indexOf(" ") + 1, request.length());
+		int id = Integer.parseInt(idStr);
+		if (id < 1) {
+			throw new ServiceException("incorrect id");
+		}
+		Book book = new Book();
+
 		try {
 			bookDAO.editBook(book);
 		} catch (DAOException e) {
@@ -74,12 +94,18 @@ public class LibraryServiceImpl implements LibraryService {
 	}
 
 	@Override
-	public void changeBookStatusServices(String request) throws ServiceException {
-		String idStr = request.substring(request.indexOf(" ") + 1, request.length());	
-		int id = Integer.parseInt(idStr);		
+	public void changeBookStatusService(String request) throws ServiceException {
+
 		BookDAO bookDAO = new SQLBookDao();
+		String[] splitRequest = RequestParserUtil.parseRequest(request, 2);
+
 		try {
-			bookDAO.changeBookStatus(BookStatus.NONRECOVERABLE, id);
+			int id = Integer.parseInt(splitRequest[1]);
+			int bookCount = bookDAO.rowCount();
+			if (id < 1 || bookCount < id) {
+				throw new ServiceException("incorrect id");
+			}
+			bookDAO.changeBookStatus(BookStatus.WRITTEN_OFF, id);
 		} catch (DAOException e) {
 			e.printStackTrace();
 			throw new ServiceException();
@@ -88,14 +114,19 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public void deleteBookService(String request) throws ServiceException {
-		int id = Integer.parseInt(request.substring(request.indexOf(" ") + 1, request.length()));
+
 		BookDAO bookDAO = new SQLBookDao();
+		int id = Integer.parseInt(request.substring(request.indexOf(" ") + 1, request.length()));
+
 		try {
+			int bookCount = bookDAO.rowCount();
+			if (id < 1 || bookCount < id) {
+				throw new ServiceException("incorrect id");
+			}
 			bookDAO.deleteBook(id);
 		} catch (DAOException e) {
 			throw new ServiceException();
 		}
-
 	}
 
 }
